move enum stop_type to server-task
